Shuffled around volume related repo apis to the volume repo rather than being in series.
