+++ API/Data/Repositories/SeriesRepository.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using API.Comparators;
@@ -43,12 +42,6 @@ namespace API.Data.Repositories
             _context.Series.Remove(series);
         }
 
-
-        public async Task<Series> GetSeriesByNameAsync(string name)
-        {
-            return await _context.Series.SingleOrDefaultAsync(x => x.Name == name);
-        }
-
         public async Task<bool> DoesSeriesNameExistInLibrary(string name)
         {
             var libraries = _context.Series
@@ -60,11 +53,6 @@ namespace API.Data.Repositories
                 .AsNoTracking()
                 .Where(s => libraries.Contains(s.LibraryId) && s.Name == name)
                 .CountAsync() > 1;
-        }
-
-        public Series GetSeriesByName(string name)
-        {
-            return _context.Series.SingleOrDefault(x => x.Name == name);
         }
 
         public async Task<IEnumerable<Series>> GetSeriesForLibraryIdAsync(int libraryId)
@@ -137,41 +125,12 @@ namespace API.Data.Repositories
                 .ToListAsync();
         }
 
-        public async Task<IEnumerable<VolumeDto>> GetVolumesDtoAsync(int seriesId, int userId)
-        {
-            var volumes =  await _context.Volume
-                .Where(vol => vol.SeriesId == seriesId)
-                .Include(vol => vol.Chapters)
-                .OrderBy(volume => volume.Number)
-                .ProjectTo<VolumeDto>(_mapper.ConfigurationProvider)
-                .AsNoTracking()
-                .ToListAsync();
-
-            await AddVolumeModifiers(userId, volumes);
-            SortSpecialChapters(volumes);
-
-            return volumes;
-        }
-
-        private static void SortSpecialChapters(IEnumerable<VolumeDto> volumes)
-        {
-            var sorter = new NaturalSortComparer();
-            foreach (var v in volumes.Where(vDto => vDto.Number == 0))
-            {
-                v.Chapters = v.Chapters.OrderBy(x => x.Range, sorter).ToList();
-            }
-        }
 
 
-        public async Task<IEnumerable<Volume>> GetVolumes(int seriesId)
-        {
-            return await _context.Volume
-                .Where(vol => vol.SeriesId == seriesId)
-                .Include(vol => vol.Chapters)
-                .ThenInclude(c => c.Files)
-                .OrderBy(vol => vol.Number)
-                .ToListAsync();
-        }
+
+
+
+
 
         public async Task<SeriesDto> GetSeriesDtoByIdAsync(int seriesId, int userId)
         {
@@ -185,55 +144,8 @@ namespace API.Data.Repositories
             return seriesList[0];
         }
 
-        public async Task<Volume> GetVolumeAsync(int volumeId)
-        {
-            return await _context.Volume
-                .Include(vol => vol.Chapters)
-                .ThenInclude(c => c.Files)
-                .SingleOrDefaultAsync(vol => vol.Id == volumeId);
-        }
 
-        public async Task<VolumeDto> GetVolumeDtoAsync(int volumeId)
-        {
-            return await _context.Volume
-                .Where(vol => vol.Id == volumeId)
-                .AsNoTracking()
-                .ProjectTo<VolumeDto>(_mapper.ConfigurationProvider)
-                .SingleAsync();
 
-        }
-
-        public async Task<VolumeDto> GetVolumeDtoAsync(int volumeId, int userId)
-        {
-            var volume = await _context.Volume
-                .Where(vol => vol.Id == volumeId)
-                .Include(vol => vol.Chapters)
-                .ThenInclude(c => c.Files)
-                .ProjectTo<VolumeDto>(_mapper.ConfigurationProvider)
-                .SingleAsync(vol => vol.Id == volumeId);
-
-            var volumeList = new List<VolumeDto>() {volume};
-            await AddVolumeModifiers(userId, volumeList);
-
-            return volumeList[0];
-        }
-
-        /// <summary>
-        /// Returns all volumes that contain a seriesId in passed array.
-        /// </summary>
-        /// <param name="seriesIds"></param>
-        /// <returns></returns>
-        public async Task<IEnumerable<Volume>> GetVolumesForSeriesAsync(IList<int> seriesIds, bool includeChapters = false)
-        {
-            var query = _context.Volume
-                .Where(v => seriesIds.Contains(v.SeriesId));
-
-            if (includeChapters)
-            {
-                query = query.Include(v => v.Chapters);
-            }
-            return await query.ToListAsync();
-        }
 
         public async Task<bool> DeleteSeriesAsync(int seriesId)
         {
@@ -243,10 +155,6 @@ namespace API.Data.Repositories
             return await _context.SaveChangesAsync() > 0;
         }
 
-        public async Task<Volume> GetVolumeByIdAsync(int volumeId)
-        {
-            return await _context.Volume.SingleOrDefaultAsync(x => x.Id == volumeId);
-        }
 
         /// <summary>
         /// Returns Volumes, Metadata, and Collection Tags
@@ -283,7 +191,7 @@ namespace API.Data.Repositories
         }
 
         /// <summary>
-        /// This returns a list of tuples<chapterId, seriesId> back for each series id passed
+        /// This returns a dictonary mapping seriesId -> list of chapters back for each series id passed
         /// </summary>
         /// <param name="seriesIds"></param>
         /// <returns></returns>
@@ -340,24 +248,7 @@ namespace API.Data.Repositories
                 .SingleOrDefaultAsync();
         }
 
-        private async Task AddVolumeModifiers(int userId, IReadOnlyCollection<VolumeDto> volumes)
-        {
-            var volIds = volumes.Select(s => s.Id);
-            var userProgress = await _context.AppUserProgresses
-                .Where(p => p.AppUserId == userId && volIds.Contains(p.VolumeId))
-                .AsNoTracking()
-                .ToListAsync();
 
-            foreach (var v in volumes)
-            {
-                foreach (var c in v.Chapters)
-                {
-                    c.PagesRead = userProgress.Where(p => p.ChapterId == c.Id).Sum(p => p.PagesRead);
-                }
-
-                v.PagesRead = userProgress.Where(p => p.VolumeId == v.Id).Sum(p => p.PagesRead);
-            }
-        }
 
         /// <summary>
         /// Returns a list of Series that were added, ordered by Created desc
